Create mysql function to translate ascii characters to Turkish
This function is used to convert the ascii characters corresponding to the Turkish characters in the strings held in the ascii MySQL database to Turkish characters. It is used to translate ascii equivalents of <,>, ",", &, and space characters besides Ş, ş, Ç, ç, İ, ı, Ö, ö, Ü, ü,
You can change and use it by yourself. I hope it helps.